DeepViz.jl

DeepViz.apply_colormap

DeepViz.generate_image

DeepViz.image_to_arr

DeepViz.load_image

DeepViz.positive_negative_saliency

DeepViz.save_gradient_images

DeepViz.save_image

DeepViz.set_training_false

DeepViz.set_training_true

DeepViz.stepped_images

DeepViz.viz_backprop

DeepViz.viz_integrated_gradient

DeepViz.apply_colormap — Method

apple_colormap(org_img, activation_map; map_alpha=0.4, process_img=true)

Creates a rainbow based colormap for the image, which is then applied onto org_img(original image) with different transparency to generate the combined heatmap with the org_img (original image). Function returns a tuple of combined colormap and activation map alone based heatmap.

DeepViz.generate_image — Method

 generate_image(x; resize_original=false, custom_size=Nothing)

Save the image as an RGB image defined the by the array x. If resize_original is set to be true, then the image can be resized to the custom_size. If not provided explicitly, then the image is resized to the original_size of the image saved while loading the image or the default one(ie. (224, 224))

DeepViz.image_to_arr — Method

image_to_arr(img; preprocess = true)

Convert an RGB type image to an array. If preprocess is set as true, subtract the mean from the image array.

DeepViz.load_image — Function

load_image(path, resize = false; size_save = true)

Function to load the images from the specified path. Returns as an array of image. If resize is set as true, resize the image to shape (224, 224).

DeepViz.positive_negative_saliency — Method

positive_negative_saliency(gradient)

Returns a tuple of positive and negative saliency maps based on the provided gradients.

DeepViz.save_gradient_images — Method

save_gradient_images(gradient, file_name; gray=false)

Saves the gradient as an RGB image by default. If gray is set to true, gradients are sasved as a gray image with the provided filename.

DeepViz.save_image — Method

save_image(path, x)

Save the image defined by x to the specified path.

DeepViz.set_training_false — Method

set_training_false()

Function to set the model into testmode, so that the output result do not vary while evaluating gradients(or other results) for the models which contain layers like BatchNormalization, Dropout etc.

DeepViz.set_training_true — Method

set_training_true()

Function to revert back the model into trainmode, once the gradients(or other results) are evaluated.

DeepViz.stepped_images — Method

stepped_images(img, steps)

Utility function for viz_integrated_gradient. Generate scaled images, ie. images multiplied by a factor i defined in the range of 0 to 1.0 over equal intervals of 1/steps.

DeepViz.viz_backprop — Method

viz_backprop(img, model; top_k=1, target_based=false, target_index)

Returns gradient generated with Vanilla Backpropagation for top_k classes. It’s the original saliency map algorithm for supervised deep learning from Simonyan et. al. (2013). These gradients can further be visualized using utilty functions.

Arguments

img: source image for which gradients are to be calculated

model: model to be used for visualization

top_k=1: gradients are returned for top_k classes as per predictions by the model.

target_based=false: if set to true, function returns the maximizing gradient for the specified target_index class along with gradient for top_k classes.

target_index: defines the target class, for which the maximizing gradients are to be calculated. Required when target_based is set to true.

The function returns an of gradients for top_k=1 predicted classes by default. If target_based is set to true and target_index is specified, the function returns a tuple: (gradients_for_top_k_classes, gradient_for_target_class).

DeepViz.viz_integrated_gradient — Method

viz_integrated_gradient(img, model, steps; target_based=false, target_index)

Produces gradients generated with integrated gradients from the image scaled over steps. This functions uses the viz_backprop for the calculation of gradients for each of scaled image and then returns integrated gradients as the average over all the scaled images.

Arguments

img: source image on which the gradient are to be calculated.

model: model to be used for visualization

steps: number of steps between 0 to 1.0 for which the image is to be scaled while calculating integrated gradients.

target_based=false: if set true, the integrated gradients are evaluated for the specified target_index.

target_index: index of the target class for which the maximized gradients are to be evaluated. Necessary condition when target_based is set as true.