Utility Functions

This section provide documentation reference to various Utility functions available with the DeepViz package. These utility function allows loading of image, changing image to array, saving the gradients as heatmaps and more.

DeepViz.image_to_arr — Function

image_to_arr(img; preprocess = true)

Convert an RGB type image to an array. If preprocess is set as true, subtract the mean from the image array.

DeepViz.load_image — Function

load_image(path, resize = false; size_save = true)

Function to load the images from the specified path. Returns as an array of image. If resize is set as true, resize the image to shape (224, 224).

DeepViz.generate_image — Function

 generate_image(x; resize_original=false, custom_size=Nothing)

Save the image as an RGB image defined the by the array x. If resize_original is set to be true, then the image can be resized to the custom_size. If not provided explicitly, then the image is resized to the original_size of the image saved while loading the image or the default one(ie. (224, 224))

DeepViz.save_image — Function

save_image(path, x)

Save the image defined by x to the specified path.

DeepViz.save_gradient_images — Function

save_gradient_images(gradient, file_name; gray=false)

Saves the gradient as an RGB image by default. If gray is set to true, gradients are sasved as a gray image with the provided filename.

DeepViz.positive_negative_saliency — Function

positive_negative_saliency(gradient)

Returns a tuple of positive and negative saliency maps based on the provided gradients.

DeepViz.save_gradcam — Function

save_gradcam(gradient, original_image_path, grad_file_name, heatmap_file_name, combined_file_name)

Function to save the gradient calculated using gradcam as a gray image, gradient heatmap and the combined heatmap with different transparency.

Arguments

gradient: Gradient calculated using the viz_gradcam function

original_image_path: Path to the original image file, for which the gradient is calculated.

grad_file_name: Filename with path, to which the gradient image is to be saved as a gray image

heatmap_file_name: Filename to save the gradient heatmap

combined_file_name: Filename to save the gradient the combined heatmap with original image.

DeepViz.apply_colormap — Function

apple_colormap(org_img, activation_map; map_alpha=0.4, process_img=true)

Creates a rainbow based colormap for the image, which is then applied onto org_img(original image) with different transparency to generate the combined heatmap with the org_img (original image). Function returns a tuple of combined colormap and activation map alone based heatmap.

DeepViz.set_training_false — Function

set_training_false()

Function to set the model into testmode, so that the output result do not vary while evaluating gradients(or other results) for the models which contain layers like BatchNormalization, Dropout etc.

DeepViz.set_training_true — Function

set_training_true()

Function to revert back the model into trainmode, once the gradients(or other results) are evaluated.