DeepViz.jl

DeepViz.ImageGenerator

DeepViz.ImageGenerator

DeepViz.apply_colormap

DeepViz.generate_image

DeepViz.grad_times_image

DeepViz.image_to_arr

DeepViz.load_image

DeepViz.positive_negative_saliency

DeepViz.save_gradcam

DeepViz.save_gradient_images

DeepViz.save_image

DeepViz.set_training_false

DeepViz.set_training_true

DeepViz.smooth_grad

DeepViz.stepped_images

DeepViz.viz_backprop

DeepViz.viz_deconvolution

DeepViz.viz_gradcam

DeepViz.viz_guidedbackprop

DeepViz.viz_guidedgradcam

DeepViz.viz_integrated_gradient

DeepViz.ImageGenerator — Method

ImageGenerator(model, target_class)

Function to initialize the ImageGenerator for the model and target_class. Returns a struct with fields model, img and target. Here, the image is randomly generated which is then to be further updated.

DeepViz.ImageGenerator — Method

ImageGenerator(save_dir; img_name = "generated", niters::Int = 150, lr = 6.0)

Generates the image for which the prediction probability is maximized for a target class after niter iterations using backpropagation. Before using this function, the ImageGenerator must be initialized using ImageGenerator(model, target_class).

Arguments

save_dir: Directory to which the generated image is to be saved

img_name="generated": Image name with which file is to be saved withoout file extension

niters::Int=150: number of iterations for which the generated image is updated with backpropagation

lr=6.0: Learning rate to be used while updating the generated image with respect to backpropagated gradients to maximize the specific class.

DeepViz.apply_colormap — Method

apple_colormap(org_img, activation_map; map_alpha=0.4, process_img=true)

Creates a rainbow based colormap for the image, which is then applied onto org_img(original image) with different transparency to generate the combined heatmap with the org_img (original image). Function returns a tuple of combined colormap and activation map alone based heatmap.

DeepViz.generate_image — Method

 generate_image(x; resize_original=false, custom_size=Nothing)

Save the image as an RGB image defined the by the array x. If resize_original is set to be true, then the image can be resized to the custom_size. If not provided explicitly, then the image is resized to the original_size of the image saved while loading the image or the default one(ie. (224, 224))

DeepViz.grad_times_image — Method

grad_times_image(img, model; top_k=1)

An extension to the saliency approach. Returns an array of products of the img with the gradient evaluated with respect to the provided img for the model, for top_k prediction classes.

DeepViz.image_to_arr — Method

image_to_arr(img; preprocess = true)

Convert an RGB type image to an array. If preprocess is set as true, subtract the mean from the image array.

DeepViz.load_image — Function

load_image(path, resize = false; size_save = true)

Function to load the images from the specified path. Returns as an array of image. If resize is set as true, resize the image to shape (224, 224).

DeepViz.positive_negative_saliency — Method

positive_negative_saliency(gradient)

Returns a tuple of positive and negative saliency maps based on the provided gradients.

DeepViz.save_gradcam — Method

save_gradcam(gradient, original_image_path, grad_file_name, heatmap_file_name, combined_file_name)

Function to save the gradient calculated using gradcam as a gray image, gradient heatmap and the combined heatmap with different transparency.

Arguments

gradient: Gradient calculated using the viz_gradcam function

original_image_path: Path to the original image file, for which the gradient is calculated.

grad_file_name: Filename with path, to which the gradient image is to be saved as a gray image

heatmap_file_name: Filename to save the gradient heatmap

combined_file_name: Filename to save the gradient the combined heatmap with original image.

DeepViz.save_gradient_images — Method

save_gradient_images(gradient, file_name; gray=false)

Saves the gradient as an RGB image by default. If gray is set to true, gradients are sasved as a gray image with the provided filename.

DeepViz.save_image — Method

save_image(path, x)

Save the image defined by x to the specified path.

DeepViz.set_training_false — Method

set_training_false()

Function to set the model into testmode, so that the output result do not vary while evaluating gradients(or other results) for the models which contain layers like BatchNormalization, Dropout etc.

DeepViz.set_training_true — Method

set_training_true()

Function to revert back the model into trainmode, once the gradients(or other results) are evaluated.

DeepViz.smooth_grad — Method

smooth_grad(img, model; method=viz_backprop,top_k=1, mean=0, var=1, n_imgs=50, normalized_n_scaled=true)

Method to visually sharpen the gradient based saliency maps. The key idea is to evaluate the average of gradients of images obtained after adding gaussian noise to the original image(img). The results suggest the estimated smoothed gradient, leads to visually more coherent sensitivity maps than the unsmoothed gradient, with the resulting visualization aligning better to the human eye with meaningful features.

Arguments

img: Original image of interest for visualization

model: Model to be used for visualization

method=viz_backprop: Method to be used while evaluating gradients. Can take methods like viz_guidedbackprop, viz_backprop, viz_deconvolution or any other custom function.

top_k=1: function evaluates the smoothed gradient for top_k predictions just as in case of normal viz_backprop function.

mean=0: Mean of the gaussian noise to be added to the image to generate sample space

var=1: Variance of the gaussian noise to be added to image to generate sample space

n_imgs=50: Number of images to be generated and used in sample space while evaluation of smooth_grad.

normalized_n_scaled=true: if true, then the noise is first scaled to 0-255 before adding to the image.

Ref: (https://arxiv.org/pdf/1706.03825.pdf)

DeepViz.stepped_images — Method

stepped_images(img, steps)

Utility function for viz_integrated_gradient. Generate scaled images, ie. images multiplied by a factor i defined in the range of 0 to 1.0 over equal intervals of 1/steps.

DeepViz.viz_backprop — Method

viz_backprop(img, model; top_k=1, target_based=false, target_index)

Returns gradient generated with Vanilla Backpropagation for top_k classes. It’s the original saliency map algorithm for supervised deep learning from Simonyan et. al. (2013). These gradients can further be visualized using utilty functions.

Arguments

img: source image for which gradients are to be calculated

model: model to be used for visualization

top_k=1: gradients are returned for top_k classes as per predictions by the model.

target_based=false: if set to true, function returns the maximizing gradient for the specified target_index class along with gradient for top_k classes.

target_index: defines the target class, for which the maximizing gradients are to be calculated. Required when target_based is set to true.

The function returns an of gradients for top_k=1 predicted classes by default. If target_based is set to true and target_index is specified, the function returns a tuple: (gradients_for_top_k_classes, gradient_for_target_class).

DeepViz.viz_deconvolution — Method

viz_deconvolution(img, model; top_k=1, target_based=false, target_index=-1)

In this method of Gradient visualization, the backpropagation of RELU activation function are overridden (so called Guided RELU) in such a way that only the non-negative gradients are backpropagated. Here, the guided RELU is applied onto the output gradients which stands as the main difference from the viz_guidedbackprop. Once the activation function is overridden, the gradient is evaluated here using the viz_backprop function.

Note: The function currently doesn't support ResNet type models. See also: viz_backprop, viz_guidedbackprop

DeepViz.viz_gradcam — Method

viz_gradcam(img, model, layer; top_k=1)

Layer attribution method that computes the gradients of the target output with respect to the given layer, averages for each output channel (dimension 2 of output), and multiplies the average gradient for each channel by the layer activations. The results are summed over all channels.

It returns an array of tuples of (gradCAM, prediction_probability, class_index) for top_k predicted classes.

DeepViz.viz_guidedbackprop — Method

viz_guidedbackprop(img, model; top_k=1, target_based=false, target_index=-1)

In this method of Gradient visualization, the backpropagation of RELU activation function are overridden (so called Guided RELU) in such a way that only the non-negative gradients are backpropagated. Here, the guided RELU is applied onto the input gradients which stands as the main difference from the viz_deconvolution. Once the activation function is overridden, the gradient is evaluated here using the viz_backprop function.

Note: The function currently doesn't support ResNet type models. See also: viz_backprop, viz_deconvolution

DeepViz.viz_guidedgradcam — Method

viz_guidedgradcam(img, model, layer; top_k=1)

It computes the element-wise product of guided backpropagation attributions (evaluated using viz_guidedbackprop) with resized (layer) GradCAM attributions (computed using viz_gradcam). GradCAM attributions are computed with respect to a given layer, and attributions are resized to match the input size.

It returns an array of tuples of (guidedgradcam, prediction_probability, class_index) for top_k predicted classes.

Also see: viz_guidedbackprop, viz_guidedgradcam.

DeepViz.viz_integrated_gradient — Method

viz_integrated_gradient(img, model, steps; target_based=false, target_index)

Produces gradients generated with integrated gradients from the image scaled over steps. This functions uses the viz_backprop for the calculation of gradients for each of scaled image and then returns integrated gradients as the average over all the scaled images.

Arguments

img: source image on which the gradient are to be calculated.

model: model to be used for visualization

steps: number of steps between 0 to 1.0 for which the image is to be scaled while calculating integrated gradients.

target_based=false: if set true, the integrated gradients are evaluated for the specified target_index.

target_index: index of the target class for which the maximized gradients are to be evaluated. Necessary condition when target_based is set as true.